refactor article service and controller

+++ services/article.js
@@ -1,4 +1,5 @@
 const articleRepository = require("../repository/article");
+const _throw = require("../utils/throw");
 
 exports.findAll = async () => {
   const result = await articleRepository.findAll();
@@ -19,29 +20,62 @@ exports.findOneById = async (id) => {
 };
 
 exports.addArticle = async (body) => {
-  const { title, content, userId, themeId } = body;
+  const { title, content, themeId, user } = body;
+  const { id: userId, admin } = user;
+
+  if (!admin) return { codeStatus: 403, msg: "Insufficient right !" };
+
   const createdAt = new Date();
 
-  return await articleRepository.addArticle(
-    title,
-    content,
-    userId,
-    themeId,
-    createdAt
-  );
+  try {
+    const result = await articleRepository.addArticle(
+      title,
+      content,
+      userId,
+      themeId,
+      createdAt
+    );
+
+    if (result.rowCount === 0) _throw("Echec created");
+
+    return { codeStatus: 201, msg: "Created successfully" };
+  } catch (error) {
+    return { codeStatus: 400, msg: error.message };
+  }
 };
 
 exports.updateArticle = async (articleId, body) => {
-  const { title, content, themeId } = body;
+  const { title, content, themeId, user } = body;
+  const { admin } = user;
 
-  return await articleRepository.updateArticle(
-    articleId,
-    title,
-    content,
-    themeId
-  );
+  if (!admin) return { codeStatus: 403, msg: "Insufficient right !" };
+
+  try {
+    const result = await articleRepository.updateArticle(
+      articleId,
+      title,
+      content,
+      themeId
+    );
+
+    if (result.rowCount === 0) _throw("Updated failed");
+
+    return { codeStatus: 200, msg: "Updated successfully" };
+  } catch (error) {
+    return { codeStatus: 400, msg: error.message };
+  }
 };
 
-exports.deleteArticle = async (id) => {
-  return await articleRepository.deleteArticle(id);
+exports.deleteArticle = async (id, admin) => {
+  if (!admin) return { codeStatus: 403, msg: "Insufficient right !" };
+
+  try {
+    const result = await articleRepository.deleteArticle(id);
+
+    if (result.rowCount === 0) _throw("Deleted failed");
+
+    return { codeStatus: 200, msg: "Deleted successfully" };
+  } catch (error) {
+    return { codeStatus: 400, msg: error.message };
+  }
 };
